Move exception handlers to `:infrastructure:async`
`:infrastructure:execeptions` had a dependency on `:infrastructure:async` that
would cause a circular dependency in some situations. For example, if
`:infrastructure:metrics` included `:infrastructure:exceptions`:

```
Circular dependency between the following tasks:
:infrastructure:async:compileJava
\--- :infrastructure:metrics:compileJava
     \--- :infrastructure:exceptions:compileJava
          \--- :infrastructure:async:compileJava (*)
```

To resolve this problem, this commit moves the two exception handlers that used
`async.SafeFuture` from `:infrastructure:execeptions:ExceptionUtil` to
`:infrastructure:async:ExceptionHandlers`.

Also, the `:infrastructure:async` dependency has been removed from
`:infrastructure:execeptions` as it is no longer required.
